Moved notepad recycler classes to base module.
