+++ base/notepad/build.gradle
@@ -2,6 +2,7 @@ plugins {
     id 'com.android.library'
     id 'kotlin-android'
     id 'kotlin-kapt'
+    id 'kotlin-android-extensions'
 }
 
 apply from: rootProject.file('dependencies.gradle')
@@ -19,6 +20,8 @@ android {
         targetCompatibility JavaVersion.VERSION_1_8
     }
     kotlinOptions {
+        // Because of error More than one file was found with OS independent path 'META-INF/notepad_debug.kotlin_module'.
+        freeCompilerArgs = ["-module-name", project.path.substring(1).replace(':', '-')]
         jvmTarget = '1.8'
     }
     sourceSets {
@@ -26,12 +29,18 @@ android {
         test.java.srcDirs += 'src/test/kotlin'
         androidTest.java.srcDirs += 'src/androidTest/kotlin'
     }
+    androidExtensions {
+        experimental = true
+    }
 }
 
 dependencies {
+    implementation project(":core")
+
     implementation(
             deps.kotlin.stdlib,
-            deps.dagger.dagger
+            deps.dagger.dagger,
+            deps.material.material
     )
     kapt(
             deps.dagger.processor
