+++ base/notepad/src/main/kotlin/com/dvinc/base/notepad/presentation/adapter/notepad/NotepadDiffUtilsCallback.kt
@@ -3,7 +3,7 @@
  * All rights reserved.
  */
 
-package com.dvinc.notepad.presentation.adapter.notepad
+package com.dvinc.base.notepad.presentation.adapter.notepad
 
 import androidx.recyclerview.widget.DiffUtil
 import com.dvinc.base.notepad.presentation.model.NoteUi
