+++ base/notepad/src/main/kotlin/com/dvinc/base/notepad/presentation/adapter/notepad/NotepadSwipeCallback.kt
@@ -3,7 +3,7 @@
  * All rights reserved.
  */
 
-package com.dvinc.notepad.presentation.adapter.notepad
+package com.dvinc.base.notepad.presentation.adapter.notepad
 
 import android.graphics.Canvas
 import android.graphics.Color
