+++ base/notepad/src/main/kotlin/com/dvinc/base/notepad/presentation/adapter/notepad/NotepadSwipeDirection.kt
@@ -1,4 +1,4 @@
-package com.dvinc.notepad.presentation.adapter.notepad
+package com.dvinc.base.notepad.presentation.adapter.notepad
 
 enum class NotepadSwipeDirection {
