+++ base/notepad/src/main/res/layout/item_note.xml
