+++ features/notepad/src/main/kotlin/com/dvinc/notepad/ui/notepad/NotepadViewModel.kt
@@ -1,4 +1,4 @@
-package com.dvinc.notepad.presentation.ui.notepad
+package com.dvinc.notepad.ui.notepad
 
 import androidx.lifecycle.viewModelScope
 import com.dvinc.core.extension.safeLaunch
@@ -7,8 +7,9 @@ import com.dvinc.core.ui.BaseViewModel
 import com.dvinc.notepad.R
 import com.dvinc.notepad.common.DEFAULT_NOTE_ID
 import com.dvinc.notepad.domain.usecase.notepad.NotepadUseCase
-import com.dvinc.notepad.presentation.adapter.notepad.NotepadSwipeDirection
+import com.dvinc.base.notepad.presentation.adapter.notepad.NotepadSwipeDirection
 import com.dvinc.base.notepad.presentation.mapper.NotePresentationMapper
+import com.dvinc.notepad.presentation.ui.notepad.NotepadFragmentDirections
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
