+++ features/notepad/src/main/res/navigation/nav_graph.xml
@@ -6,7 +6,7 @@
 
     <fragment
         android:id="@+id/notepadFragment"
-        android:name="com.dvinc.notepad.presentation.ui.notepad.NotepadFragment"
+        android:name="com.dvinc.notepad.ui.notepad.NotepadFragment"
         android:label="NotepadFragment"
         tools:layout="@layout/fragment_notepad">
         <action
@@ -23,7 +23,7 @@
 
     <fragment
         android:id="@+id/noteFragment"
-        android:name="com.dvinc.notepad.presentation.ui.note.NoteFragment"
+        android:name="com.dvinc.notepad.ui.note.NoteFragment"
         android:label="NoteFragment"
         tools:layout="@layout/fragment_note">
 
@@ -35,7 +35,7 @@
 
     <fragment
         android:id="@+id/archiveFragment"
-        android:name="com.dvinc.notepad.presentation.ui.archive.ArchiveFragment"
+        android:name="com.dvinc.notepad.ui.archive.ArchiveFragment"
         android:label="ArchiveFragment"
         tools:layout="@layout/fragment_archive">
     </fragment>
